# shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sms/": {
      "get": {
        "tags": [
          "sms"
        ],
        "summary": "List Sms",
        "operationId": "list_sms",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "phone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Phone"
            }
          },
          {
            "name": "starred",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Starred"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sms/stats": {
      "get": {
        "tags": [
          "sms"
        ],
        "summary": "Sms Stats",
        "operationId": "sms_stats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/sms/phones": {
      "get": {
        "tags": [
          "sms"
        ],
        "summary": "Sms Phones",
        "operationId": "sms_phones",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response Sms-Sms Phones"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/sms/{sms_id}": {
      "get": {
        "tags": [
          "sms"
        ],
        "summary": "Get Sms",
        "operationId": "get_sms",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sms_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sms Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sms"
        ],
        "summary": "Delete Sms",
        "operationId": "delete_sms",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sms_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sms Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sms/upload": {
      "post": {
        "tags": [
          "sms"
        ],
        "summary": "Upload Sms",
        "operationId": "upload_sms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/sms/upload/batch": {
      "post": {
        "tags": [
          "sms"
        ],
        "summary": "Batch Upload Sms",
        "operationId": "batch_upload_sms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsBatchUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsBatchUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/sms/{sms_id}/star": {
      "patch": {
        "tags": [
          "sms"
        ],
        "summary": "Star Sms",
        "operationId": "star_sms",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sms_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sms Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsStarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sms/batch-star": {
      "post": {
        "tags": [
          "sms"
        ],
        "summary": "Batch Star Sms",
        "operationId": "batch_star_sms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsBatchStarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsBatchActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/sms/batch-delete": {
      "post": {
        "tags": [
          "sms"
        ],
        "summary": "Batch Delete Sms",
        "operationId": "batch_delete_sms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsBatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsBatchActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/v1/sms/inbound": {
      "post": {
        "tags": [
          "sms-forward"
        ],
        "summary": "Sms Inbound",
        "operationId": "sms_inbound",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "X-Sms-Forward-Version",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sms-Forward-Version"
            }
          },
          {
            "name": "X-Sms-Forward-Message-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sms-Forward-Message-Id"
            }
          },
          {
            "name": "X-Sms-Forward-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sms-Forward-Device-Id"
            }
          },
          {
            "name": "X-Sms-Forward-Rule-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sms-Forward-Rule-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsForwardInboundRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SmsBatchActionResponse": {
        "properties": {
          "updated": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated"
          },
          "deleted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted"
          }
        },
        "type": "object",
        "title": "SmsBatchActionResponse"
      },
      "SmsBatchDeleteRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "SmsBatchDeleteRequest"
      },
      "SmsBatchStarRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Ids"
          },
          "starred": {
            "type": "boolean",
            "title": "Starred"
          }
        },
        "type": "object",
        "required": [
          "ids",
          "starred"
        ],
        "title": "SmsBatchStarRequest"
      },
      "SmsBatchUploadRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/SmsUploadRequest"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "SmsBatchUploadRequest"
      },
      "SmsBatchUploadResponse": {
        "properties": {
          "success": {
            "type": "integer",
            "title": "Success"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "success",
          "failed",
          "ids"
        ],
        "title": "SmsBatchUploadResponse"
      },
      "SmsForwardDevice": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "androidSdk": {
            "type": "integer",
            "title": "Androidsdk"
          },
          "appVersion": {
            "type": "string",
            "title": "Appversion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "model",
          "manufacturer",
          "androidSdk",
          "appVersion"
        ],
        "title": "SmsForwardDevice"
      },
      "SmsForwardInboundRequest": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "device": {
            "$ref": "#/components/schemas/SmsForwardDevice"
          },
          "rule": {
            "$ref": "#/components/schemas/SmsForwardRule"
          },
          "message": {
            "$ref": "#/components/schemas/SmsForwardMessage"
          },
          "meta": {
            "$ref": "#/components/schemas/SmsForwardMeta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event",
          "version",
          "device",
          "rule",
          "message",
          "meta"
        ],
        "title": "SmsForwardInboundRequest"
      },
      "SmsForwardMessage": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          },
          "subscriptionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscriptionid"
          },
          "simSlot": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simslot"
          },
          "partCount": {
            "type": "integer",
            "title": "Partcount",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "from",
          "body",
          "timestamp"
        ],
        "title": "SmsForwardMessage"
      },
      "SmsForwardMeta": {
        "properties": {
          "receivedAt": {
            "type": "integer",
            "title": "Receivedat"
          },
          "sentAt": {
            "type": "integer",
            "title": "Sentat"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt",
            "default": 1
          },
          "contentLength": {
            "type": "integer",
            "title": "Contentlength"
          },
          "contentSha256": {
            "type": "string",
            "title": "Contentsha256"
          }
        },
        "type": "object",
        "required": [
          "receivedAt",
          "sentAt",
          "contentLength",
          "contentSha256"
        ],
        "title": "SmsForwardMeta"
      },
      "SmsForwardRule": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "senderFilter": {
            "type": "string",
            "title": "Senderfilter"
          }
        },
        "type": "object",
        "required": [
          "id",
          "senderFilter"
        ],
        "title": "SmsForwardRule"
      },
      "SmsListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SmsRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "items"
        ],
        "title": "SmsListResponse"
      },
      "SmsRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "starred": {
            "type": "boolean",
            "title": "Starred"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "phone",
          "content",
          "received_at",
          "starred",
          "created_at"
        ],
        "title": "SmsRead"
      },
      "SmsStarRequest": {
        "properties": {
          "starred": {
            "type": "boolean",
            "title": "Starred"
          }
        },
        "type": "object",
        "required": [
          "starred"
        ],
        "title": "SmsStarRequest"
      },
      "SmsStatsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "starred": {
            "type": "integer",
            "title": "Starred"
          },
          "today": {
            "type": "integer",
            "title": "Today"
          },
          "this_week": {
            "type": "integer",
            "title": "This Week"
          },
          "this_month": {
            "type": "integer",
            "title": "This Month"
          }
        },
        "type": "object",
        "required": [
          "total",
          "starred",
          "today",
          "this_week",
          "this_month"
        ],
        "title": "SmsStatsResponse"
      },
      "SmsUploadRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "content"
        ],
        "title": "SmsUploadRequest"
      },
      "SmsUploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SmsUploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}